Moved table dir under ClusterJ
