+++ rondb/src/main/java/site/ycsb/db/clusterj/table/package-info.java
@@ -18,5 +18,5 @@
 /**
  * YCSB binding for <a href="https://rondb.com/">RonDB</a>.
  */
-package site.ycsb.db.table;
+package site.ycsb.db.clusterj.table;
