+++ rondb/src/main/java/site/ycsb/db/rest/RonDBRestClient.java
@@ -31,6 +31,7 @@ import site.ycsb.Client;
 import site.ycsb.Status;
 import site.ycsb.db.RonDBClient;
 import site.ycsb.db.clusterj.RonDBConnection;
+import site.ycsb.db.clusterj.table.UserTableHelper;
 import site.ycsb.db.rest.ds.BatchRequest;
 import site.ycsb.db.rest.ds.BatchResponse;
 import site.ycsb.db.rest.ds.BatchSubOperation;
@@ -39,7 +40,6 @@ import site.ycsb.db.rest.ds.MyHttpClientSync;
 import site.ycsb.db.rest.ds.MyHttpClient;
 import site.ycsb.db.rest.ds.PKRequest;
 import site.ycsb.db.rest.ds.PKResponse;
-import site.ycsb.db.table.UserTableHelper;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
